+++ R/environments.R
@@ -32,50 +32,42 @@ print.mark_environments <- function(x, ...) {
 #' @return
 #' * `ls_all()`, `objects_all()`: A named list for each of the environments in the `search()` path with all the objects found in that environment
 ls_all <- function(all.names = FALSE) {
-  sapply(search(), ls_switch, all.names = all.names)
+  sapply(search(), function(s) ls(as.environment(s), all.names = all.names))
 }
 
 #'@export
 #'@rdname list_environments
 objects_all <- ls_all
 
-ls_switch <- function(x, all.names = FALSE) {
-  switch(
-    x,
-    R_GlobalEnv = ls(globalenv(), all.names = all.names),
-    R_EmptyEnv = "",
-    base = ls(baseenv(), all.names = all.names),
-    ls(as.environment(x), all.names = all.names)
-  )
-}
-
 
 #' List Objects - extensions
 #'
 #' @inheritParams base::ls
 #' @return A `character` vector of names
-#' @export
 #' @name ls_ext
+NULL
 
-#' @export
-#' @rdname ls_ext
-ls_dataframe <- function(pattern, all.names = FALSE) {
-  do_ls(is.data.frame, pattern = pattern, all.names = all.names)
+make_do_ls <- function(FUN) {
+  FUN <- match.fun(FUN)
+  function(pattern, all.names = FALSE, envir = parent.frame()) {
+    do_ls(FUN, pattern = pattern, all.names = all.names, envir = envir)
+  }
 }
 
 #' @export
 #' @rdname ls_ext
-ls_function <- function(pattern, all.names = FALSE) {
-  do_ls(is.function, pattern = pattern, all.names = all.names)
-}
+ls_dataframe <- make_do_ls(is.data.frame)
 
 #' @export
 #' @rdname ls_ext
-ls_object <- function(pattern, all.names = FALSE) {
-  do_ls(is.object, pattern = pattern, all.names = all.names)
-}
+ls_function <- make_do_ls(is.function)
 
-do_ls <- function(FUN, pattern, all.names = FALSE) {
-  .ls <- ls(envir = parent.frame(3), pattern = pattern)
-  .ls[vap_lgl(.ls, function(x) FUN(get0(x)))]
+#' @export
+#' @rdname ls_ext
+ls_object <- make_do_ls(is.object)
+
+
+do_ls <- function(FUN, pattern, all.names = FALSE, envir = parent.frame()) {
+  .ls <- ls(envir = envir, pattern = pattern, all.names = all.names)
+  .ls[vap_lgl(.ls, function(x) FUN(get0(x, envir = envir)))]
 }