+++ tests/testthat/test-nas.R
@@ -3,6 +3,10 @@ test_that("remove_na()", {
   expect_equal(remove_na(x), c(1, 2, 3))
   expect_equal(remove_na(as.list(x)), list(1, 2, numeric(), 3, numeric()))
   expect_error(remove_na(data.frame(x = 1)))
+
+  res <- remove_na(fact(x))
+  exp <- struct(1:4, c("fact", "factor"), levels = c("1", "2", "3", "NaN"))
+  expect_identical(res, exp)
 })
 
 test_that("remove_null()", {
@@ -57,7 +61,7 @@ test_that("tableNA() works", {
     struct(
       c(`TRUE` = 0L, `FALSE` = 3L),
       dim = 2L,
-      dimnames = set_names0(list(c("TRUE", "FALSE")), ""),
+      dimnames = set_names0(list(c("TRUE", "FALSE")), "x"),
       class = "table"
     )
   )
@@ -87,4 +91,15 @@ test_that("tableNA() works", {
       class = "table"
     )
   )
+})
+
+test_that("omit_na() works", {
+  x <- letters[1:4]
+  res <- omit_na(x)
+  exp <- struct(x, "character", na = integer(), valid = 1:4)
+  expect_identical(res, exp)
+
+  x[2] <- NA
+  res <- omit_na(x)
+  exp <- struct(x[c(1, 3, 4)], "character", na = 2L, valid = c(1L, 3L, 4L))
 })